+++ Projects/STM32F446ZE-Nucleo/Applications/USB_Device/DFU_Standalone/Src/usbd_conf.c
@@ -433,7 +433,7 @@ USBD_StatusTypeDef USBD_LL_SetUSBAddress(USBD_HandleTypeDef * pdev,
   */
 USBD_StatusTypeDef USBD_LL_Transmit(USBD_HandleTypeDef * pdev,
                                     uint8_t ep_addr,
-                                    uint8_t * pbuf, uint16_t size)
+                                    uint8_t * pbuf, uint32_t size)
 {
   HAL_PCD_EP_Transmit(pdev->pData, ep_addr, pbuf, size);
   return USBD_OK;
@@ -449,7 +449,7 @@ USBD_StatusTypeDef USBD_LL_Transmit(USBD_HandleTypeDef * pdev,
   */
 USBD_StatusTypeDef USBD_LL_PrepareReceive(USBD_HandleTypeDef * pdev,
                                           uint8_t ep_addr,
-                                          uint8_t * pbuf, uint16_t size)
+                                          uint8_t * pbuf, uint32_t size)
 {
   HAL_PCD_EP_Receive(pdev->pData, ep_addr, pbuf, size);
   return USBD_OK;
